More logger changes for the bootstrap deployer
The logger in charmhub client was incorrect because it requires
lots of shims all the time. Instead if we just break out the
logger and the logger factory, we can now break the dependency.
